Map And Overall Quality Of Life Changes
The Map Button On The Nav Bar Now Takes The User To A Map Web Page Which Used OpenStreetView API to Display The Users Current Location I Attempted To Allow For The Map To Show Local "Cex" Stores However This Dose Not Work Yet. More Quality Of Life Changes The Form For Creating A New Game Has Been Changed To Look Better.